config from configmanager
